refactor: HWndCtrlFactory
